+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add polling to unlock accounts after a configurable cooldown period (AUTHLIB-180)
+
 ## [4.3.0] - 2026-09-16
 
 ### Added